melhorias de otimização parte 2
